moved add/remove axiom and save methods to Ontology class
